requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Association Rules Mining
mlxtend>=0.22.0

# Semi-supervised Learning
scikit-learn>=1.3.0
label-studio>=1.7.0

# Clustering & Evaluation
yellowbrick>=1.5

# Notebook & Automation
jupyter>=1.0.0
notebook>=6.5.0
papermill>=2.4.0
nbconvert>=7.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0

# Optional: GUI/Demo
streamlit>=1.28.0
gradio>=3.50.0
